# docs/openapi/openapi.yaml
openapi: 3.0.0
info:
  title: Call Calendar API
  version: 0.0.0
...
paths:
  /api/v1/bookings/{bookingId}:
    get:
      operationId: Bookings_getBooking
      parameters:
        - name: bookingId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: The request has succeeded.
          content:
            application/json:
              schema:
                anyOf:
                  - $ref: '#/components/schemas/Booking'
                  - $ref: '#/components/schemas/ErrorResponse'
  /api/v1/bookings/{bookingId}/cancel:
    post:
      operationId: Bookings_cancelBooking
      parameters:
        - name: bookingId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: The request has succeeded.
          content:
            application/json:
              schema:
                anyOf:
                  - $ref: '#/components/schemas/Booking'
                  - $ref: '#/components/schemas/ErrorResponse'
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CancelBookingRequest'
  /api/v1/bookings/{bookingId}/reschedule:
    post:
      operationId: Bookings_rescheduleBooking
      parameters:
        - name: bookingId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: The request has succeeded.
          content:
            application/json:
              schema:
                anyOf:
                  - $ref: '#/components/schemas/Booking'
                  - $ref: '#/components/schemas/ErrorResponse'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RescheduleBookingRequest'
  /api/v1/hosts/{slug}/availability:
    get:
      operationId: Availability_getAvailability
      parameters:
        - name: slug
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: The request has succeeded.
          content:
            application/json:
              schema:
                anyOf:
                  - $ref: '#/components/schemas/AvailabilitySettings'
                  - $ref: '#/components/schemas/ErrorResponse'
    put:
      operationId: Availability_updateAvailability
      parameters:
        - name: slug
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: The request has succeeded.
          content:
            application/json:
              schema:
                anyOf:
                  - $ref: '#/components/schemas/AvailabilitySettings'
                  - $ref: '#/components/schemas/ErrorResponse'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateAvailabilityRequest'
  /api/v1/hosts/{slug}/bookings:
    get:
      operationId: HostBookings_listHostBookings
      parameters:
        - name: slug
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: The request has succeeded.
          content:
            application/json:
              schema:
                anyOf:
                  - type: array
                    items:
                      $ref: '#/components/schemas/Booking'
                  - $ref: '#/components/schemas/ErrorResponse'
    post:
      operationId: HostBookings_createBooking
      parameters:
        - name: slug
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: The request has succeeded.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '201':
          description: The request has succeeded and a new resource has been created as a result.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Booking'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateBookingRequest'
  /api/v1/hosts/{slug}/event-types:
    get:
      operationId: EventTypes_listEventTypes
      parameters:
        - name: slug
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: The request has succeeded.
          content:
            application/json:
              schema:
                anyOf:
                  - type: array
                    items:
                      $ref: '#/components/schemas/EventType'
                  - $ref: '#/components/schemas/ErrorResponse'
    post:
      operationId: EventTypes_createEventType
      parameters:
        - name: slug
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: The request has succeeded.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '201':
          description: The request has succeeded and a new resource has been created as a result.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EventType'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateEventTypeRequest'
  /api/v1/hosts/{slug}/event-types/{eventTypeId}:
    patch:
      operationId: EventTypes_updateEventType
      parameters:
        - name: slug
          in: path
          required: true
          schema:
            type: string
        - name: eventTypeId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: The request has succeeded.
          content:
            application/json:
              schema:
                anyOf:
                  - $ref: '#/components/schemas/EventType'
                  - $ref: '#/components/schemas/ErrorResponse'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateEventTypeRequest'
    delete:
      operationId: EventTypes_deleteEventType
      parameters:
        - name: slug
          in: path
          required: true
          schema:
            type: string
        - name: eventTypeId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: The request has succeeded.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '204':
          description: 'There is no content to send for this request, but the headers may be useful. '
  /api/v1/hosts/{slug}/settings:
    get:
      operationId: ApiV1_getHostSettings
      parameters:
        - name: slug
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: The request has succeeded.
          content:
            application/json:
              schema:
                anyOf:
                  - $ref: '#/components/schemas/HostSettings'
                  - $ref: '#/components/schemas/ErrorResponse'
  /api/v1/hosts/{slug}/slots:
    get:
      operationId: ApiV1_listSlots
      parameters:
        - name: slug
          in: path
          required: true
          schema:
            type: string
        - name: date
          in: query
          required: false
          schema:
            $ref: '#/components/schemas/LocalDate'
          explode: false
        - name: eventTypeId
          in: query
          required: false
          schema:
            type: string
          explode: false
      responses:
        '200':
          description: The request has succeeded.
          content:
            application/json:
              schema:
                anyOf:
                  - $ref: '#/components/schemas/AvailabilityDay'
                  - $ref: '#/components/schemas/ErrorResponse'
components:
  schemas:
    ApiError:
      type: object
      required:
        - code
        - message
      properties:
        code:
          allOf:
            - $ref: '#/components/schemas/ErrorCode'
          description: Код ошибки для программной обработки.
        message:
          type: string
          description: Человекочитаемое сообщение.
        details:
          type: object
          additionalProperties: {}
          description: Дополнительные детали (например, ошибки полей).
      description: Единая модель ошибки для всех не-2xx ответов.
    AvailabilityDay:
      type: object
      required:
        - timeZone
        - slots
      properties:
        date:
          type: string
          allOf:
            - $ref: '#/components/schemas/LocalDate'
          nullable: true
          description: Дата фильтра, если была запрошена.
        timeZone:
          type: string
          description: IANA-пояс, к которому относятся startAt слотов.
        slots:
          type: array
          items:
            $ref: '#/components/schemas/Slot'
...
    AvailabilitySettings:
      type: object
      required:
        - timeZone
        - slotDurationMin
        - bufferMin
        - minNoticeMin
        - horizonDays
        - ranges
      properties:
        timeZone:
          type: string
        slotDurationMin:
          type: integer
          format: int32
        bufferMin:
          type: integer
          format: int32
        minNoticeMin:
          type: integer
          format: int32
        horizonDays:
          type: integer
          format: int32
        ranges:
          type: array
          items:
            $ref: '#/components/schemas/AvailabilityRange'
    Booking:
      type: object
      required:
        - id
        - hostSlug
        - eventTypeId
        - startAt
        - endAt
        - timeZone
        - clientName
        - clientEmail
        - status
        - createdAt
      properties:
        id:
          type: string
          description: UUID — публичный идентификатор для ссылок отмены и переноса.
        hostSlug:
          type: string
        eventTypeId:
          type: string
        startAt:
          allOf:
            - $ref: '#/components/schemas/UtcDateTime'
          description: Начало встречи (UTC ISO 8601).
        endAt:
          $ref: '#/components/schemas/UtcDateTime'
        timeZone:
          type: string
        clientName:
          type: string
        clientEmail:
          type: string
        clientPhone:
          type: string
          nullable: true
        clientNotes:
          type: string
          nullable: true
        status:
          $ref: '#/components/schemas/BookingStatus'
        cancellationReason:
          type: string
          nullable: true
          description: Причина отмены, если бронь отменена.
        createdAt:
          type: string
    BookingStatus:
      type: string
      enum:
        - confirmed
        - cancelled
    CancelBookingRequest:
      type: object
      properties:
        reason:
          type: string
          description: Необязательная причина отмены.
    CreateBookingRequest:
      type: object
      required:
        - eventTypeId
        - startAt
        - clientName
        - clientEmail
      properties:
        eventTypeId:
          type: string
        startAt:
          allOf:
            - $ref: '#/components/schemas/UtcDateTime'
          description: Желаемое начало слота (UTC ISO 8601).
        clientName:
          type: string
        clientEmail:
          type: string
        clientPhone:
          type: string
        clientNotes:
          type: string
    CreateEventTypeRequest:
      type: object
      required:
        - slug
        - title
        - durationMin
        - locationType
      properties:
        slug:
          type: string
        title:
          type: string
        description:
          type: string
        durationMin:
          type: integer
          format: int32
        locationType:
          $ref: '#/components/schemas/LocationType'
        isActive:
          type: boolean
...
    ErrorResponse:
      type: object
      required:
        - error
      properties:
        error:
          $ref: '#/components/schemas/ApiError'
      description: 'Тело ответа с ошибкой: конверт `{ error: ApiError }`.'
    EventType:
      type: object
      required:
        - id
        - slug
        - title
        - durationMin
        - locationType
        - isActive
      properties:
        id:
          type: string
        slug:
          type: string
        title:
          type: string
        description:
          type: string
          nullable: true
        durationMin:
          type: integer
          format: int32
          description: 'Длительность встречи в минутах (MVP: 30).'
        locationType:
          $ref: '#/components/schemas/LocationType'
        isActive:
          type: boolean
          description: Активен ли тип для записи гостем.
    HostSettings:
      type: object
      required:
        - slug
        - name
        - timeZone
      properties:
        slug:
          type: string
        name:
          type: string
        timeZone:
          type: string
          description: IANA-пояс хоста, например Europe/Moscow.
      description: Публичные настройки организатора по slug.
    LocalDate:
      type: string
      format: date
      description: Локальная дата (YYYY-MM-DD) в поясе хоста.
    LocationType:
      type: string
      enum:
        - online
        - offline
        - phone
    RescheduleBookingRequest:
      type: object
      required:
        - startAt
      properties:
        startAt:
          allOf:
            - $ref: '#/components/schemas/UtcDateTime'
          description: Новое начало слота (UTC ISO 8601).
...
    UpdateAvailabilityRequest:
      type: object
      properties:
        timeZone:
          type: string
        slotDurationMin:
          type: integer
          format: int32
        bufferMin:
          type: integer
          format: int32
        minNoticeMin:
          type: integer
          format: int32
        horizonDays:
          type: integer
          format: int32
        ranges:
          type: array
          items:
            $ref: '#/components/schemas/AvailabilityRange'
    UpdateEventTypeRequest:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
        durationMin:
          type: integer
          format: int32
        locationType:
          $ref: '#/components/schemas/LocationType'
        isActive:
          type: boolean
    UtcDateTime:
      type: string
      format: date-time
      description: Момент времени в UTC, ISO 8601 с миллисекундами и суффиксом `Z`; для отображения — поле timeZone (IANA).
servers:
  - url: http://localhost:3000
    description: Local development
    variables: {}